feat: Optimize Mapbox line layer style and zoom level
This commit optimizes the Mapbox line layer style and zoom level in the HomeScreen component. It updates the line color to #000000 and reduces the line width to 5. Additionally, it sets the default zoom level to 15 in the MapBoxComponent. These changes improve the visual appearance and performance of the map display in the application.